backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
requests>=2.32.0
Pillow>=10.4.0
beautifulsoup4>=4.12.3
cos-python-sdk-v5>=1.9.25
edge-tts>=6.1.0
